requirements.txt
streamlit
pandas
python-pptx
openai>=1.0.0
python-docx
tiktoken
